requirements.txt
numpy
scipy
scikit-learn
librosa
soundfile
streamlit
joblib
streamlit-webrtc
google-genai